Kill me :3
Right, for some inexplicable reason, the bleedin' input element is black, and it absolutely refuses to turn white no matter what I try! I've spent hours fiddlin' with the CSS, trying every bloody combination of color properties, background properties, and nothin' works! It's like the input element is determined to stay black just to spite me. Why is this so hard? It should be a simple change! Yet, here I am, battlin' an input element that seems to have a mind of its own. It needs to be white so it contrasts well on the black background, but apparently, that's too much to ask for. I’ve even sacrificed my sanity at the altar of web design, and still, the bleedin' input element remains black. Unbelievable!